# add index to improve vote query performance