A delegate noticed that a db query was taking over 7 seconds to execute on their forging node:

```sql
UPDATE mem_accounts m SET vote = (SELECT COALESCE(SUM(b.balance), 0) as vote FROM mem_accounts2delegates a, mem_accounts b where a."accountId" = b.address AND a."dependentId" = encode(m."publicKey", 'hex')) WHERE m."isDelegate" = 1;
```

I took a look at the query explain plan and noticed it was missing an important index that was causing the query to perform poorly for the volume of data on the mainnet. Adding this index improves performance for this query significantly.

Without this index the network could become unstable since the query execution time (>5 sec is approaching the block time (8 sec). 

I put the query plans before and after the fix as a comment in the sql migration script. Adding this one index 

Before: 2801.446 ms
After: 120.902 ms