@boldninja 2017-12-04
@bradyo can you provide me with your ARK address?